+++ app/archives/landing-old/head.tsx
@@ -1,0 +1,8 @@
+export default function Head() {
+    return (
+        <>
+            <title>Home | ky28059.github.io</title>
+            <meta name="description" content="Personal portfolio website and technology testing chamber." />
+        </>
+    )
+}
