+++ app/archives/landing-old/page.tsx
@@ -1,27 +1,12 @@
-import {ReactNode, useEffect, useState} from 'react';
-import Head from 'next/head';
-import Link from 'next/link';
-
-// Icons
-import {BsChevronCompactDown, BsDiscord} from 'react-icons/bs';
-import {FiArrowRight, FiLink2} from 'react-icons/fi';
-import {BiSun} from 'react-icons/bi';
-import {MdExpandLess, MdExpandMore} from 'react-icons/md';
-import {GoMarkGithub} from 'react-icons/go';
-
-// Utilities
-import {langToColor} from '../../components/Project';
+import Project from './Project';
+import Header from './Header';
+import {BsChevronCompactDown} from 'react-icons/bs';
 
 
 export default function Home() {
     return (
         <div className="dark:text-white bg-gray-100 dark:bg-gradient-to-r dark:from-black dark:to-gray-900 selection:bg-blue-500 selection:text-white">
-            <Head>
-                <title>Home | ky28059.github.io</title>
-                <meta name="description" content="Personal portfolio website and technology testing chamber." />
-            </Head>
-
-            <ArchivedHeader />
+            <Header />
 
             <section className="bg-white dark:bg-gray-900 pt-14 pb-2 mb-10">
                 <div className="text-center py-4">
@@ -38,7 +23,7 @@ export default function Home() {
             <section>
                 <h1 className="text-5xl font-bold text-center mb-8">Projects</h1>
                 <div className="container mx-auto flex justify-center flex-wrap gap-6">
-                    <ArchivedProject
+                    <Project
                         name="RBot"
                         img="/projects/RBot.png"
                         langs={['ts']}
@@ -46,8 +31,8 @@ export default function Home() {
                         discord="https://discord.com/oauth2/authorize?client_id=684587440777986090&scope=bot&permissions=8"
                     >
                         Discord mod bot using djs 13 and sequelize.
-                    </ArchivedProject>
-                    <ArchivedProject
+                    </Project>
+                    <Project
                         name="WATT"
                         img="/projects/watt-white.png"
                         langs={['ts', 'html', 'scss']}
@@ -55,7 +40,7 @@ export default function Home() {
                         web="https://gunnwatt.web.app/"
                     >
                         Gunn schedule app and UGWA successor.
-                    </ArchivedProject>
+                    </Project>
                     {/*
                     <Project
                         name="Word Generator"
@@ -64,7 +49,7 @@ export default function Home() {
                         web="https://ky28059.github.io/#/word-gen"
                     />
                     */}
-                    <ArchivedProject
+                    <Project
                         name="Geckobot"
                         img="/projects/GeckoBot.png"
                         langs={['c#', 'q#']}
@@ -72,8 +57,8 @@ export default function Home() {
                         discord="https://discord.com/oauth2/authorize?client_id=766064505079726140&scope=bot&permissions=379968"
                     >
                         Discord.net bot with Drive API and quantum computing.
-                    </ArchivedProject>
-                    <ArchivedProject
+                    </Project>
+                    <Project
                         name="PortalBot"
                         img="/projects/PortalBot.jpg"
                         langs={['ts']}
@@ -81,8 +66,8 @@ export default function Home() {
                         discord="https://discord.com/oauth2/authorize?client_id=827738852902043698&scope=bot&permissions=8"
                     >
                         TypeScript djs bot for real time cross server communication.
-                    </ArchivedProject>
-                    <ArchivedProject
+                    </Project>
+                    <Project
                         name="GunnHacks 7.0"
                         img="/projects/gunnhacks.png"
                         langs={['js', 'html', 'scss']}
@@ -90,8 +75,8 @@ export default function Home() {
                         web="https://7.0.gunnhacks.com"
                     >
                         Website for GunnHacks 7.0 (2020-2021).
-                    </ArchivedProject>
-                    <ArchivedProject
+                    </Project>
+                    <Project
                         name="GunnHacks 8.0"
                         img="/projects/gunnhacks.png"
                         langs={['svelte', 'html', 'css']}
@@ -99,37 +84,37 @@ export default function Home() {
                         web="https://gunnhacks.com"
                     >
                         SvelteKit website for GunnHacks 8.0.
-                    </ArchivedProject>
-                    <ArchivedProject
+                    </Project>
+                    <Project
                         name="Guava Bot"
                         img="/projects/Guava.png"
                         langs={['ts']}
                         gh="https://github.com/ky28059/guava-bot"
                     >
                         <code>!whois</code> bot for an elusive discord server.
-                    </ArchivedProject>
-                    <ArchivedProject
+                    </Project>
+                    <Project
                         name="Guava Bot Slack"
                         img="/projects/Guava.png"
                         langs={['ts']}
                         gh="https://github.com/ky28059/guava-bot-slack"
                     >
                         Slack bot for an elusive slack workspace.
-                    </ArchivedProject>
-                    <ArchivedProject
+                    </Project>
+                    <Project
                         name="hoco-radio-2021"
                         langs={['py']}
                         gh="https://github.com/ky28059/hoco-radio-2021"
                     >
                         Python script for the Raspberry Pi radio in the Junior float.
-                    </ArchivedProject>
-                    <ArchivedProject
+                    </Project>
+                    <Project
                         name="AP CSA"
                         langs={['rkt']}
                         gh="https://github.com/ky28059/AP-CSA"
                     >
                         Repository of assignments from the first (scheme) semester of Gunn's AP CSA.
-                    </ArchivedProject>
+                    </Project>
                 </div>
                 {/*
                 <Link href="/all"><a className="text-2xl mx-auto">More <FiArrowRight /></a></Link>
@@ -147,99 +132,5 @@ export default function Home() {
                 <br/>
             </section>
         </div>
-    )
-}
-
-function ArchivedHeader() {
-    const [open, setOpen] = useState(true);
-
-    const [scroll, setScroll] = useState(0);
-    const toggleTheme = () => document.documentElement.classList.toggle('dark');
-
-    useEffect(() => {
-        setScroll(window.scrollY);
-        document.onscroll = () => setScroll(window.scrollY);
-    }, []);
-
-    const ExpandIcon = open ? MdExpandLess : MdExpandMore;
-
-    return (
-        <header className={`flex sticky top-0 p-4 bg-white dark:bg-gray-800 text-black dark:text-white bg-opacity-90 ${scroll > 0 ? 'shadow-md' : 'hover:shadow-md'} backdrop-blur-sm z-50 h-20 transition-shadow duration-300 ease-in-out ${!open ? 'h-10' : ''}`}>
-            <div className={`flex flex-grow items-center h-full ${!open ? 'hidden' : ''}`}>
-                <a href="https://github.com/ky28059">
-                    <img src="/pfp.png" alt="Profile" className="rounded-full" height={50} width={50}/>
-                </a>
-                <h1 className="text-3xl font-bold ml-4 mr-8">ky28059</h1>
-
-                <span className="p-4">
-                    <Link href="/"><a className="text-inherit hover:no-underline">Home</a></Link>
-                </span>
-                <span className="p-4">
-                    <Link href="/concepts"><a className="text-inherit hover:no-underline">Concepts</a></Link>
-                </span>
-            </div>
-            <div className="flex items-center text-xl gap-4">
-                <BiSun className="cursor-pointer" onClick={toggleTheme}/>
-                <ExpandIcon className="cursor-pointer" onClick={() => setOpen(x => !x)} />
-            </div>
-        </header>
-    )
-}
-
-type ProjectProps = {
-    name: string, children: ReactNode, img?: string /* StaticImageData */, langs?: string[],
-    gh?: string, discord?: string, web?: string
-};
-function ArchivedProject(props: ProjectProps) {
-    const {name, children: desc, img, langs, gh, discord, web} = props;
-
-    return (
-        <div className="flex flex-col p-4 gap-3 shadow-lg bg-white dark:bg-gray-800 rounded-md transform hover:scale-110 transition transition-transform duration-150 w-64">
-            <span className="flex items-center space-x-3 mb-1.5">
-                {img && (
-                    <img
-                        src={img}
-                        className="rounded-full h-8 w-8 bg-gray-200/50 dark:bg-gray-900/50"
-                        alt={`${name}-icon`}
-                        width={40}
-                        height={40}
-                    />
-                )}
-                <h3 className="text-xl font-bold">{name}</h3>
-            </span>
-            {langs && (
-                <span className="space-x-2 ml-1 flex">
-                    {langs.map(lang => <ProjectTag lang={lang} key={lang} />)}
-                </span>
-            )}
-            <p>{desc}</p>
-            <span className="flex gap-2 mt-auto text-xl">
-                {gh && (
-                    <a href={gh} rel="noopener noreferrer" target="_blank" className="text-inherit hover:no-underline">
-                        <GoMarkGithub />
-                    </a>
-                )}
-                {discord && (
-                    <a href={discord} rel="noopener noreferrer" target="_blank" className="text-inherit hover:no-underline">
-                        <BsDiscord />
-                    </a>
-                )}
-                {web && (
-                    <a href={web} rel="noopener noreferrer" target="_blank" className="text-inherit hover:no-underline">
-                        <FiLink2 />
-                    </a>
-                )}
-            </span>
-        </div>
-    );
-}
-
-function ProjectTag(props: {lang: string}) {
-    const color = langToColor(props.lang);
-    return (
-        <div
-            className="rounded-full w-3 h-3 ring-2 ring-black ring-opacity-75 dark:ring-opacity-50"
-            style={{backgroundColor: color}}
-        />
     )
 }
